Warning Cleanups
Math: conversion to rad
Robot Consts: gyro gain
PeriodicThread: elasped time should have been float

+++ src/core/architecture/thread/PeriodicThread.hpp
@@ -40,7 +40,7 @@ namespace KSystem
 			{
 				t.restart ();
 				this->Execute();
-				double el = t.elapsed();
+				float el = t.elapsed();
 
 				if ( el  > ThreadPeriod && ThreadPeriod > 0 )
 					Logger::Instance().WriteMsg(GetName(), "Decrease Freq!!!", Logger::ExtraInfo );
